[#D3D-2689] Skips inactive/disabled/incomplete scene objects during export
Invalid can mean inactive, disables, incomplete or not found.
For skins, it means that some bones are not in the current export selection
Disabled components or gameobjects are also ignored.

A texture is invalid if it is used in scene (and visible) but the asset
doesn't exist in asset database. (no export possible)

Moreover, if not visible object remains in the selection, the export is
canceled. (only meshes are considered as visible objects for now)